+++ init_project.sh
@@ -3,15 +3,16 @@
 vagrant_dir=$PWD
 magento_ce_dir="${vagrant_dir}/magento2ce"
 magento_ee_dir="${magento_ce_dir}/magento2ee"
-config_path="${vagrant_dir}/local.config/config.yaml"
+config_path="${vagrant_dir}/etc/config.yaml"
+host_os=$(bash "${vagrant_dir}/scripts/host/get_host_os.sh")
 
 # Enable trace printing and exit on the first error
 set -ex
 
-bash "${vagrant_dir}/scripts/host/shell/check_requirements.sh"
+bash "${vagrant_dir}/scripts/host/check_requirements.sh"
 
 # Generate random IP address and host name to prevent collisions, if not specified explicitly in local config
-if [ ! -f "${vagrant_dir}/local.config/config.yaml" ]; then
+if [ ! -f "${vagrant_dir}/etc/config.yaml" ]; then
     cp "${config_path}.dist" ${config_path}
 fi
 random_ip=$(( ( RANDOM % 240 )  + 12 ))
@@ -35,12 +36,17 @@ if [ ${force_project_cleaning} -eq 1 ]; then
 fi
 
 if [ ! -d ${magento_ce_dir} ]; then
+    if [[ ${host_os} == "Windows" ]]; then
+        git config --global core.autocrlf false
+        git config --global core.eol LF
+        git config --global diff.renamelimit 5000
+    fi
     # Check out CE repository
-    repository_url_ce=$(bash "${vagrant_dir}/scripts/host/shell/get_variable_value.sh" "repository_url_ce")
+    repository_url_ce=$(bash "${vagrant_dir}/scripts/host/get_variable_value.sh" "repository_url_ce")
     git clone ${repository_url_ce} ${magento_ce_dir}
     # Check out EE repository
     # By default EE repository is not specified and EE project is not checked out
-    repository_url_ee=$(bash "${vagrant_dir}/scripts/host/shell/get_variable_value.sh" "repository_url_ee")
+    repository_url_ee=$(bash "${vagrant_dir}/scripts/host/get_variable_value.sh" "repository_url_ee")
     if [ -n "${repository_url_ee}" ]; then
         git clone ${repository_url_ee} ${magento_ee_dir}
     fi
@@ -48,26 +54,28 @@ fi
 
 var_generation_dir="${magento_ce_dir}/var/generation"
 if [ ! -d ${var_generation_dir} ] ; then
-    host_os=$(bash "${vagrant_dir}/scripts/host/shell/get_host_os.sh")
-    if [[ ${host_os} == "Windows" || $(bash "${vagrant_dir}/scripts/host/shell/get_variable_value.sh" "guest_use_nfs") == 0 ]]; then
+    if [[ ${host_os} == "Windows" || $(bash "${vagrant_dir}/scripts/host/get_variable_value.sh" "guest_use_nfs") == 0 ]]; then
         mkdir ${var_generation_dir}
     fi
 fi
 
 # Update Magento dependencies via Composer
 cd ${magento_ce_dir}
-bash "${vagrant_dir}/scripts/host/shell/composer.sh" install
+bash "${vagrant_dir}/scripts/host/composer.sh" install
 
 # Install necessary vagrant plugins if not installed
-if ! vagrant plugin list | grep -q 'hostmanager' ; then
+if ! vagrant plugin list | grep -q 'vagrant-hostmanager' ; then
     vagrant plugin install vagrant-hostmanager
 fi
-if ! vagrant plugin list | grep -q 'vbguest' ; then
+if ! vagrant plugin list | grep -q 'vagrant-vbguest' ; then
     vagrant plugin install vagrant-vbguest
+fi
+if ! vagrant plugin list | grep -q 'vagrant-host-shell' ; then
+    vagrant plugin install vagrant-host-shell
 fi
 
 # Create vagrant project
 cd ${vagrant_dir}
 vagrant up
 
-bash "${vagrant_dir}/scripts/host/shell/configure_php_storm.sh"
+bash "${vagrant_dir}/scripts/host/configure_php_storm.sh"