+++ scripts/host/get_host_os.sh
@@ -4,6 +4,9 @@ if [ "$(uname)" == "Darwin" ]; then
     echo "OSX"
 elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
     echo "Linux"
-elif [ "$(expr substr $(uname -s) 1 10)" == "MINGW32_NT" ]; then
+elif [ "$(expr substr $(uname -s) 1 5)" == "MINGW" ]; then
     echo "Windows"
+else
+    echo "Unknown host OS"
+    exit 255
 fi