+++ CHANGELOG.md
@@ -12,13 +12,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
  - Moved provisioning scripts
  - Magento project directory moved to vagrant project root. Current structure is as follows: `vagrant-magento/magento2ce/magento2ee`
- - PHP 7.0 is installed by default instead of PHP 5.5.9 (can be configured in [local.config/config.yaml.dist](local.config/config.yaml.dist))
+ - PHP 7.0 is installed by default instead of PHP 5.5.9 (can be configured in [etc/config.yaml.dist](etc/config.yaml.dist))
+ - Renamed configuration folder from `local.config` to `etc`
+ - Set minimum Vagrant version as 1.8
 
 ### Added
 
  - Implemented static value of forwarded SSH port to prevent necessity to reconfigure software accessing guest via SSH
  - Implemented collision prevention for IP address and host name (in case when several machines are created at once)
- - Added configuration file [local.config/config.yaml.dist](local.config/config.yaml.dist)
+ - Added configuration file [etc/config.yaml.dist](etc/config.yaml.dist)
  - Added PHP 7.0 support
  - Added PHP Storm configuration during project initialization
  - Added [project initialization script](init_project.sh) for *nix and OSX hosts